dtype bsz

+++ training.py
@@ -2,6 +2,7 @@ import os
 
 import numpy as np
 import pandas as pd
+import tensorflow as tf
 from tqdm.auto import tqdm
 
 import optim
@@ -48,7 +49,7 @@ def epoch_train(args, strategy, step_fn, ds):
     pbar = tqdm(ds)
     for input in pbar:
         # Step
-        step_args = input + (float(args.bsz),)
+        step_args = input + (tf.cast(args.bsz, args.dtype),)
         acc, ce_loss, con_loss = strategy.run(step_fn, args=step_args)
         acc = strategy.reduce('SUM', acc, axis=None)
         ce_loss = strategy.reduce('SUM', ce_loss, axis=None)
